+++ Home.py
@@ -4,7 +4,8 @@ import google.auth.transport.requests
 import google.oauth2.id_token
 import os 
 from logzero import logger
-import database_functions as db_funcs
+
+import helper.database_functions as db_funcs
 
 os.environ["OAUTHLIB_INSECURE_TRANSPORT"] = "1"
 os.environ['GOOGLE_CLIENT_ID'] = st.secrets['google_oauth']['client_id']
@@ -25,8 +26,6 @@ flow = Flow.from_client_config(
     redirect_uri= st.secrets["google_oauth"]["redirect_uris"][0]
 )
 
-sql_db, cursor = db_funcs.initialize_database()
-
 def google_oauth():
     authorization_url, state = flow.authorization_url(prompt='consent')
     st.session_state['state'] = state
@@ -41,71 +40,94 @@ def process_callback():
             request = google.auth.transport.requests.Request()
             id_info = google.oauth2.id_token.verify_oauth2_token(
                 credentials._id_token, request, os.environ['GOOGLE_CLIENT_ID'], clock_skew_in_seconds=3)
+            st.session_state['credentials'] = credentials_to_dict(credentials)
             return id_info
         except Exception as e:
             st.error(f"Error fetching token: {e}")
             return None
     return None
 
-if __name__ == "__main__":
-    st.markdown("# Home 📝")
-    st.sidebar.markdown("# Home 📝")
-    st.subheader("Welcome! This is a smart assistant to help break down your big seemingly complex tasks, into smaller steps that makes it easier to build on each day!")
+def credentials_to_dict(credentials):
+    return {'token': credentials.token, 'refresh_token': credentials.refresh_token, 'token_uri': credentials.token_uri, 'client_id': credentials.client_id, 'client_secret': credentials.client_secret, 'scopes': credentials.scopes}
+
+def initial_display_elements():
+    st.markdown("""
+        <style>
+            h2 {
+                color: #1E90FF; /* Blue color for h2 headings */
+            }
+            h3 {
+                color: #1E90FF; /* Blue color for h3 headings */
+            }
+            h4 {
+                color: #1E90FF; /* Blue color for h3 headings */
+            }
+            p, li {
+                color: #FFFFFF; /* White color for paragraph and list items */
+            }
+        </style>
+        """, unsafe_allow_html=True)
+    
+    st.header("TaskBreakdown 📝", divider='rainbow')
+    # st.sidebar.markdown("# Home 📝")
+    st.subheader("Break down your big goals into manageable steps")
+
+    st.markdown("""
+    ## About TaskBreakdown
+    TaskBreakdown is a smart assistant designed to help you break down your long-term goals into manageable steps. Whether you're looking to lose weight, switch jobs, or achieve any other complex goal, TaskBreakdown provides a personalized plan tailored to your needs.
+
+    #### What Can You Do with TaskBreakdown?
+    - **Set Long-Term Goals**: Define your big goals and let the assistant help you break them down.
+    - **Create Action Plans**: Get detailed, actionable steps to achieve your goals.
+    - **Sync with Google Calendar**: Schedule your tasks and stay on track by syncing them with your Google Calendar.
+    - **Track Progress**: Use the structured plan to make tangible progress on your goals.
+
+    #### Example Goals:
+    - **Lose Weight**: "How can I lose weight?"
+    - **Switch Jobs**: "How can I switch jobs in my industry but don't know where to start?"
+    - **Run a Marathon**: "How do I prepare to run a marathon in six months?"
+    - **Learn a New Skill**: "What steps should I take to learn Python programming?"
+                """)
     
     st.markdown("""
-    ## How to Use This Tool:
-    1. **Login with Google**: Use the button below to log in with your Google account.
-    2. **Enter Gemini API Key**: After logging in, enter your Gemini API key. If you don't have one, follow these steps to get it:
-       - Go to [Gemini Studio](https://aistudio.google.com/app/apikey) and sign up or log in.
-       - Navigate to the API section in your account settings.
-       - Generate a new API key and copy it.
-    3. **Navigate to the Todolist Tab**: Once your API key is saved, go to the Todolist tab to start using the assistant.
+    ## How to Begin:
+    1. **Sign up as a tester**: Send a mail or contact me on Twitter to sign up as a tester. After confirmation from me, proceed with step 2.
+    2. **Login with Google**: Use the "Login with Google" link below to authenticate your Google account with all the necessary permissions. Wait for the setup to be ready.
+    3. **Navigate to the Todolist Tab**: Go to the Todolist tab using the side bar and read the How to Use pop down to get help with using the assistant.
     """)
-     
+
+@st.cache_data(show_spinner=False)
+def get_cached_api_key(email:str):
+    db, cursor = db_funcs.initialize_database()
+    return db_funcs.get_user_api_key(cursor, email)
+
+if __name__ == "__main__":
+    initial_display_elements()
+    db, cursor = db_funcs.initialize_database()
+
     if 'user_info' not in st.session_state:
+        st.session_state['credentials'] = None
         st.session_state['user_info'] = None
-        st.session_state['gemini_api_key'] = None
-        st.markdown(''' **You're not logged in, please login to use the assistant** ''')
+        st.warning(body="You're not logged in, please login to use the assistant")
 
     if st.session_state['user_info']:
-        user_info = st.session_state['user_info']
-        email = user_info['email']
-        logger.debug(user_info)
-        st.session_state['user_id'] = db_funcs.get_user_id(cursor, user_info['email'])
-        if not st.session_state['user_id']:
-            db_funcs.save_user(cursor, sql_db, user_info['email'], user_info['name'], user_info['picture'])
-            st.session_state['user_id'] = db_funcs.get_user_id(cursor, user_info['email'])
+        if not db_funcs.is_user_present(cursor, st.session_state['user_info']['email']):
+            db_funcs.save_user(cursor, db, st.session_state['user_info']['email'], st.session_state['user_info']['name'], st.session_state['user_info']['picture'])
     
-        
-        st.write(f"Welcome {user_info['given_name']} ")
-        api_key = db_funcs.get_user_api_key(cursor, email)
-        logger.debug(f"api key = {api_key}")
-        if api_key:
-            st.session_state['gemini_api_key'] = api_key
-            st.write("API Key found and loaded")
-        else:
-            api_key_input = st.text_input("Enter your Gemini API Key", type="password")
-            if api_key_input:
-                st.session_state['gemini_api_key'] = api_key_input
-                db_funcs.save_user(cursor, sql_db, email,user_info['name'],user_info['picture'],st.session_state['gemini_api_key'])
-                api_key = db_funcs.get_user_api_key(cursor, email)
-                st.write(f"API key saved successfully{api_key}")
-                # api_key = db_funcs.get_user_api_key(cursor, email)
-
-        if st.session_state['gemini_api_key']:
-            logger.debug(f"api key in session {api_key}")
-            st.write("You can now head onto the Todolist tab, to talk to the assistant :)")
-            
+        st.info(f"Welcome {st.session_state['user_info']['name']} ")
+        st.success("Setup Ready! You can now head onto the Todolist tab, to talk to the assistant :)")
+        logger.info(f"Welcome {st.session_state['user_info']['name']} ")
     else:
         user_info = process_callback()
         if user_info:
             st.session_state['user_info'] = user_info
-            st.write(f"You've logged in as {user_info['given_name']}")
+            st.info(f"You've logged in as {user_info['name']}")
             st.rerun()
         else:
             google_oauth()
 
     st.markdown("""
     ---
-    **Need help? Contact support at [santoshramakrishnan24@gmail.com](mailto:santoshramakrishnan24@gmail.com)**
+    **Need help? Contact support at [santoshramakrishnan24@gmail.com](mailto:santoshramakrishnan24@gmail.com) \
+    Reach out to me on [Twitter](https://x.com/SantoshKutti24)**
     """)